Clean up conflicts
